+++ android/src/main/kotlin/com/firebase_auth_ui/FirebaseAuthUiPlugin.kt
@@ -118,6 +118,8 @@ class FirebaseAuthUiPlugin(private val activity: Activity) : MethodCallHandler, 
                         .build())
                 "twitter" -> providers.add(AuthUI.IdpConfig.TwitterBuilder()
                         .build())
+                "phone" -> providers.add(AuthUI.IdpConfig.PhoneBuilder()
+                        .build())
             }
         }
         return providers
@@ -127,10 +129,19 @@ class FirebaseAuthUiPlugin(private val activity: Activity) : MethodCallHandler, 
         var instance = AuthUI.getInstance()
                 .createSignInIntentBuilder()
                 .setAvailableProviders(providers)
+
+        val logoResourceId = getLogoResourceId()
+        if (logoResourceId > 0)
+            instance = instance.setLogo(logoResourceId)
+
         if (tos?.isNotEmpty() == true && privacyPolicy?.isNotEmpty() == true)
             instance = instance.setTosAndPrivacyPolicyUrls(tos, privacyPolicy)
 
         activity.startActivityForResult(instance.build(),
                 AUTH_REQUEST_CODE)
+    }
+
+    private fun getLogoResourceId(): Int {
+        return activity.resources.getIdentifier("auth_ui_logo", "drawable", activity.packageName)
     }
 }